Fix to first load not getting collection properly
BGG has some form of lazy load system for the collection API, so that it will ask you to wait for a few seconds on the first request of the collection rather than give you the collection straight away. What this means is that the collection bit wasn't working correctly. What I have done then is put a loop in here just to keep refreshing the api until it returns the collection properly