Updates to accomodate changes to the BoardGameGeek API
I have updated the code to accomodate changes to the BoardGameGeek API around authorisation for their API. You will now need to acquire an authorisation token to be able to use this script (details of how to do this can be found here: https://boardgamegeek.com/using_the_xml_api). I have also added in an option to import a users game collection.